requirements/common.txt
torch
torchvision
pyyaml
matplotlib
statistics
